# internal/adapters/inbound/http/openapi.yaml
openapi: 3.1.0
info:
  title: ObservAI API
  version: 1.0.0
  description: |
    Provider-agnostic observability analysis gateway.

    All successful responses are wrapped in `WrapperDtoResponde` with a stable
    `data` and `metadata` envelope. Provider-specific fields never leak into
    the public contract — adapters normalize them into the schemas documented
    below.

    The interaction chat answers only questions related to an active analysis.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /v1/me/password:
    post:
      tags: [profile]
      summary: Change the authenticated user's password and revoke active sessions.
      operationId: changePassword
      responses:
        "204": { description: Password updated; client must log in again. }
        "401": { description: Current password is incorrect. }
...